+++ tests/vm/cases/data_list_insert_middle_shift.ts
@@ -18,8 +18,8 @@ export class DataListInsertMiddleShift extends UdonSharpBehaviour {
     list.RemoveAt(0 as UdonInt);
 
     Debug.Log(list.Count);
-    Debug.Log(list.get_Item(0 as UdonInt).Float);
-    Debug.Log(list.get_Item(1 as UdonInt).Float);
-    Debug.Log(list.get_Item(2 as UdonInt).Float);
+    Debug.Log(list.get_Item(0 as UdonInt).Double);
+    Debug.Log(list.get_Item(1 as UdonInt).Double);
+    Debug.Log(list.get_Item(2 as UdonInt).Double);
   }
 }
